fix: grant unlock rewards for all discovery methods
