+++ common/src/main/java/com/evandev/fieldguide/network/ScanUnlockPacket.java
@@ -1,22 +1,12 @@
 package com.evandev.fieldguide.network;
 
-import com.evandev.fieldguide.Constants;
-import com.evandev.fieldguide.config.ModConfig;
 import com.evandev.fieldguide.server.ScanVerifier;
-import com.evandev.fieldguide.server.ServerFieldGuideManager;
 import com.evandev.fieldguide.server.progress.FieldGuideProgressManager;
 import com.evandev.fieldguide.server.progress.PlayerFieldGuideProgress;
-import net.minecraft.commands.CommandSource;
-import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.ArrayList;
-import java.util.List;
 
 public class ScanUnlockPacket {
     private final ResourceLocation entryId;
@@ -62,65 +52,6 @@ public class ScanUnlockPacket {
         if (progress.isUnlocked(entryId)) return;
         if (!ScanVerifier.verifyScan(player, entryId, scannedTargetId, targetBlockPos, targetEntityId)) return;
 
-        long gameTime = player.serverLevel().dayTime();
-        progress.unlock(entryId.toString(), gameTime);
-
-        ModConfig config = ModConfig.get();
-        if (config.grantXpOnScan && config.xpAmountOnScan > 0) {
-            player.giveExperiencePoints(config.xpAmountOnScan);
-        }
-
-        executeCommands(player, entryId);
-    }
-
-    private void executeCommands(ServerPlayer player, ResourceLocation entryId) {
-        ModConfig config = ModConfig.get();
-        List<String> commandsToRun = new ArrayList<>(config.globalScanCommands);
-
-        String idStr = entryId.toString();
-        if (config.entryScanCommands.containsKey(idStr)) {
-            commandsToRun.addAll(config.entryScanCommands.get(idStr));
-        }
-
-        ResourceLocation categoryId = ServerFieldGuideManager.getInstance().getCategoryForEntryId(entryId);
-        if (categoryId != null && config.categoryScanCommands.containsKey(categoryId.toString())) {
-            commandsToRun.addAll(config.categoryScanCommands.get(categoryId.toString()));
-        }
-
-        if (!commandsToRun.isEmpty()) {
-            CommandSourceStack sourceStack = createRewardSourceStack(player, entryId);
-            for (String cmd : commandsToRun) {
-                player.getServer().getCommands().performPrefixedCommand(sourceStack, cmd);
-            }
-        }
-    }
-
-    private static CommandSourceStack createRewardSourceStack(ServerPlayer player, ResourceLocation entryId) {
-        String sourceName = Constants.MOD_ID + "/" + entryId;
-        return new CommandSourceStack(
-                new CommandSource() {
-                    @Override
-                    public void sendSystemMessage(@NotNull Component component) {
-                        Constants.LOG.info("[SCAN] {}", component.getString());
-                    }
-
-                    @Override
-                    public boolean acceptsSuccess() {
-                        return true;
-                    }
-
-                    @Override
-                    public boolean acceptsFailure() {
-                        return true;
-                    }
-
-                    @Override
-                    public boolean shouldInformAdmins() {
-                        return false;
-                    }
-                },
-                player.position(), player.getRotationVector(), player.serverLevel(),
-                2, sourceName, Component.literal(sourceName), player.getServer(), player
-        );
+        progress.unlock(player, entryId);
     }
 }
